docuagent-backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-multipart
pypdf
pdfplumber
python-docx
pandas
openpyxl
xlrd
openai
requests
